Richard's changes to read spline file and mc_configs file on master and distribute to other processes and change nconf_total from real*8 to integer.
git-svn-id: svn://hydrogene.lct.jussieu.fr/champ@93 36b7402c-5635-0410-8f0e-a47a0fc0da44

+++ qmc/variables_mod.f90
@@ -75,7 +75,7 @@ module variables_mod
   real(dp)                       :: lambda = 0.3d0
 
 ! total number of configurations (walkers)
-  real(dp)   :: nconf_total = 1
+  integer   :: nconf_total = 1
 
 ! weight of walkers
   real(dp), allocatable   :: walker_weights (:)
